ref: use arrays instead of maps for attribute representation
This change stems from the observation that using arrays and indices
instead of maps and UIDs can significantly improve parsing speed and
decrease message size in many cases. A deck of slides detailing this
observation can be found in the "Documentation" repository.

It turns out that, in order to be able to parse the attributes into
ogdf::{Node,Edge}Arrays, we also need an additional field in the
attribute message that indicates the graph object type it belong to.